Organized files and deleted unneeded ones
