+++ whiteboardflow-frontend/src/Main.js
@@ -1,15 +1,13 @@
 // Main.js
 import React, { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
-import Layout from './Layout';
-import HomePage from './HomePage';
+import Layout from './components/Layout/Layout';
+import HomePage from './components/HomePage/HomePage';
 import Whiteboard from './components/Whiteboard/Whiteboard';
-import OralTest from './pages/OralTest/OralTest';
-import Settings from './components/Settings/Settings';
-import BackEndTest from './BackEndTest';
+import OralTest from './components/OralTest/OralTest';
 import Results from './components/Result/Result';
 import { auth } from './firebase';
-import { AvatarProvider } from './Avatar/AvatarContext';
+import { AvatarProvider } from './components/Avatar/AvatarContext';
 
 function Main() {
     const [user, setUser] = useState(null);
@@ -34,8 +32,6 @@ function Main() {
                         <Route index element={<HomePage user={user} />} />
                         {/* Only allow access to these routes if user is logged in */}
                         <Route path="OralTest" element={<PrivateRoute><OralTest /></PrivateRoute>} />
-                        <Route path="Settings" element={<PrivateRoute><Settings /></PrivateRoute>} />
-                        <Route path="BackEndTest" element={<PrivateRoute><BackEndTest /></PrivateRoute>} />
                         <Route path="results" element={<PrivateRoute><Results /></PrivateRoute>} />
                         <Route path="whiteboard" element={<PrivateRoute><Whiteboard /></PrivateRoute>} />
                     </Route>
