--- whiteboardflow-frontend/src/components/AIAssistant/AIAssistant.js
@@ -1,32 +1,0 @@
-import React from 'react';
-
-/**
- * AIAssistant component that provides hints and grades based on the user's input.
- * @component
- * @property {string} data - The input used by the AI assistant for processing.
- */
-const AIAssistant = ({ data }) => {
-
-  /**
-   * Provides a hint from the user's current progress on the question.
-   * @returns {string} The hint generated by the AI assistant.
-   */
-  const hint = () => {
-
-  };
-
-  /**
-   * Grades the user based on their answer.
-   * @returns {number} The grade assigned to the input data.
-   */
-  const grade = () => {
-
-  };
-
-  return (
-    <div>
-    </div>
-  );
-};
-
-export default AIAssistant;
