+++ whiteboardflow-frontend/src/components/Avatar/Character.css
