+++ whiteboardflow-frontend/src/components/Avatar/helper-working.js
