+++ whiteboardflow-frontend/src/components/Avatar/helper.css
