+++ whiteboardflow-frontend/src/components/Avatar/helper.js
