+++ whiteboardflow-frontend/src/components/HomePage/HomePage.css
