+++ whiteboardflow-frontend/src/components/Layout/Layout.css
