--- whiteboardflow-frontend/src/components/Menu/Menu.js
@@ -1,29 +1,0 @@
-import React from 'react';
-
-/**
- * Menu component that manages the visibility of a menu.
- * @component
- */
-const Menu = () => {
-  
-  /**
-   * Displays the menu.
-   */
-  const display = () => {
-
-  };
-
-  /**
-   * Hides the menu.
-   */
-  const hide = () => {
-    
-  };
-
-  return (
-    <div>
-    </div>
-  );
-};
-
-export default Menu;
