--- whiteboardflow-frontend/src/components/Profile/Profile.js
@@ -1,22 +1,0 @@
-import React from 'react';
-
-/**
- * Profile component that displays user information.
- * @component
- */
-const Profile = () => {
-
-  /**
-   * Deletes a user profile.
-   */
-  const deleteProfile = () => {
-
-  };
-
-  return (
-    <div>
-    </div>
-  );
-};
-
-export default Profile;
